Code/Tools fixes

+++ Code/Tools/ProjectManager/Source/GemCatalog/GemRequirementDelegate.cpp
@@ -35,7 +35,7 @@ namespace O3DE::ProjectManager
         CalcRects(options, fullRect, itemRect, contentRect);
 
         QFont standardFont(options.font);
-        standardFont.setPixelSize(s_fontSize);
+        standardFont.setPixelSize(static_cast<int>(s_fontSize));
         QFontMetrics standardFontMetrics(standardFont);
 
         painter->save();
@@ -55,10 +55,10 @@ namespace O3DE::ProjectManager
         QFont gemNameFont(options.font);
         const int firstColumnMaxTextWidth = s_summaryStartX - 30;
         gemName = QFontMetrics(gemNameFont).elidedText(gemName, Qt::TextElideMode::ElideRight, firstColumnMaxTextWidth);
-        gemNameFont.setPixelSize(s_gemNameFontSize);
+        gemNameFont.setPixelSize(static_cast<int>(s_gemNameFontSize));
         gemNameFont.setBold(true);
         QRect gemNameRect = GetTextRect(gemNameFont, gemName, s_gemNameFontSize);
-        gemNameRect.moveTo(contentRect.left(), contentRect.center().y() - s_gemNameFontSize);
+        gemNameRect.moveTo(contentRect.left(), contentRect.center().y() - static_cast<int>(s_gemNameFontSize));
 
         painter->setFont(gemNameFont);
         painter->setPen(m_textColor);
